refactor: consolidate funcs for unit tests
