+++ test/src/unit/controllers/projects.controller.test.js
@@ -14,7 +14,9 @@ const { mockReq, mockRes } = require('sinon-express-mock');
 
 const projectsController = rewire('../../../../src/pfe/portal/controllers/projects.controller');
 const ProjectList = require('../../../../src/pfe/portal/modules/ProjectList');
+
 const { suppressLogOutput } = require('../../../modules/log.service');
+const { exampleProject } = require('../../../modules/projectCreation.service');
 
 chai.should();
 
@@ -97,86 +99,3 @@ describe('projects.controller.js', () => {
     });
 });
 
-const exampleProject = {
-    projectID: '0e018aa0-1515-11ea-9063-896c05b47bde',
-    name: 'node3',
-    codewindVersion: 'latest',
-    language: 'nodejs',
-    locOnDisk: '/Users/richard.waller@ibm.com/codewind-workspace/node3',
-    host: '172.18.0.5',
-    ports: {
-        exposedPort: '32771',
-        internalPort: '3000',
-        internalDebugPort: '9229',
-    },
-    workspace: '/codewind-workspace/',
-    directory: 'node3',
-    infLockFlag: false,
-    loadTestPath: '/codewind-workspace/node3/load-test',
-    projectType: 'nodejs',
-    buildLogPath: '/codewind-workspace/.logs/node3-0e018aa0-1515-11ea-9063-896c05b47bde/docker.build.log',
-    state: 'open',
-    autoBuild: true,
-    injectMetrics: 'true',
-    buildStatus: 'success',
-    detailedBuildStatus: ' ',
-    projectWatchStateId: 'ea72093628ff9aa1904c2a8de89bf009',
-    ignoredPaths: [
-        '/.project',
-        '/run-dev',
-        '/run-debug',
-        '/package-lock.json*',
-        '/nodejs_restclient.log',
-        '/nodejs_dc.log',
-        '/manifest.yml',
-        '/idt.js',
-        '/cli-config.yml',
-        '/README.md',
-        '/Jenkinsfile',
-        '/.cfignore',
-        '/load-test*',
-        '*/node_modules*',
-        '*/.git/*',
-        '*/.DS_Store',
-        '*/*.swp',
-        '*/*.swx',
-        '*/4913',
-        '*/.dockerignore',
-        '*/.gitignore',
-        '*/*~',
-        '/.settings',
-        '/chart',
-    ],
-    appStatus: 'started',
-    detailedAppStatus: {
-        severity: 'INFO',
-        message: 'Application started.',
-        notify: false,
-    },
-    lastbuild: 1575543700071,
-    appImageLastBuild: '1575543689000',
-    location: '/codewind-workspace/node3',
-    validationStatus: 'success',
-    validationResults: [],
-    isHttps: false,
-    containerId: '63d21ae7e283372c0ac40d90517b651e3cc52fdb0bbefc15c0321197d5e21f0b',
-    logs: {
-        build: [
-            {
-                origin: 'workspace',
-                files: [
-                    '/codewind-workspace/.logs/node3-0e018aa0-1515-11ea-9063-896c05b47bde/docker.build.log',
-                ],
-            },
-        ],
-        app: [
-            {
-                origin: 'workspace',
-                files: [
-                    '/codewind-workspace/.logs/node3-0e018aa0-1515-11ea-9063-896c05b47bde/app.log',
-                ],
-            },
-        ],
-    },
-    metricsAvailable: true,
-};
